+++ SIGpi_installer.sh
@@ -5,7 +5,7 @@
 ###
 
 ###
-###  REVISION: 20230107-2300
+###  REVISION: 20230129-2300
 ###
 
 ###
@@ -27,8 +27,9 @@ SIGPI_HOME=$SIGPI_ROOT/SIGpi
 SIGPI_ETC=$SIGPI_ROOT/etc
 SIGPI_SCRIPTS=$SIGPI_HOME/scripts
 SIGPI_PACKAGES=$SIGPI_HOME/packages
+SIGPI_DEBS=$SIGPI_HOME/debs
 
-# SigPi Install Support files
+# SIGpi Install Support files
 SIGPI_INSTALLER=$SIGPI_ETC/INSTALL_CONFIG
 SIGPI_CONFIG=$SIGPI_ETC/INSTALLED
 SIGPI_PKGLIST=$SIGPI_PACKAGES/PACKAGES
@@ -56,8 +57,8 @@ DESKTOP_FILES=/usr/share/applications
 DESKTOP_ICONS=/usr/share/icons
 DESKTOP_XDG_MENU=/usr/share/extra-xdg-menus
 
-# SigPi Menu category
-SIGPI_MENU_CATEGORY=SigPi
+# SIGpi Menu category
+SIGPI_MENU_CATEGORY=SIGpi
 HAMRADIO_MENU_CATEGORY=HamRadio
 
 ###
@@ -158,16 +159,18 @@ select_startscreen(){
 
 select_devices() {
     FUN=$(whiptail --title "SIGpi Installer" --clear --checklist --separate-output \
-        "Select additional devices to install " 20 80 12 \
-        "limesuite" "LimeSDR" OFF \
-        "ettus" "Ettus Research USRP UHD" OFF \
+        "Choose additional SDR devices " 20 80 12 \
+        "bladerf" "bladeRF " OFF \
+        "ettus" "Ettus Research USRP UHD " OFF \
+        "limesuite" "LimeSDR " OFF \
+        "sdrplay" "SDRPlay " OFF \
         "rfm95w" "(RPi only) Adafruit LoRa Radio Bonnet - RFM95W @ 915 MHz " OFF \
         3>&1 1>&2 2>&3)
     RET=$?
     if [ $RET -eq 1 ]; then
         $FUN = "NONE"
     fi
-    ##echo $FUN >> $SIGPI_INSTALLER
+
     IFS=' '     # space is set as delimiter
     read -ra ADDR <<< "$FUN"   # str is read into an array as tokens separated by IFS
     for i in "${ADDR[@]}"; do   # access each element of array
@@ -176,9 +179,10 @@ select_devices() {
 }
 
 select_sdrapps() {
-    FUN=$(whiptail --title "SigPi Installer" --clear --checklist --separate-output \
-        "Select SDR Applications" 20 80 12 \
+    FUN=$(whiptail --title "SIGpi Installer" --clear --checklist --separate-output \
+        "Choose general purpose SDR Applications" 20 80 12 \
         "cubicsdr" "SDR Receiver 0.2.4" OFF \
+        "gnuradio" "GNU Radio 3.10" OFF \
         "sdrangel" "SDRangel 7.8.5" OFF \
 		"sdrpp" "SDR++ 1.1.0" OFF \
         3>&1 1>&2 2>&3)
@@ -186,7 +190,7 @@ select_sdrapps() {
     if [ $RET -eq 1 ]; then
         $FUN = "NONE"
     fi
-    ##echo $FUN >> $SIGPI_INSTALLER
+
     IFS=' '     # space is set as delimiter
     read -ra ADDR <<< "$FUN"   # str is read into an array as tokens separated by IFS
     for i in "${ADDR[@]}"; do   # access each element of array
@@ -195,19 +199,21 @@ select_sdrapps() {
 }
 
 select_amateurradio() {
-    FUN=$(whiptail --title "SigPi Installer" --clear --checklist --separate-output \
-        "Select Amateur Radio Applications" 24 120 12 \
+    FUN=$(whiptail --title "SIGpi Installer" --clear --checklist --separate-output \
+        "Choose Amateur Radio Applications" 24 120 12 \
         "hamlib" "Ham Radio Control Libraries 4.5.3 " OFF \
         "fldigi" "Fldigi 4.1.18 for MFSK, PSK31, CW, RTTY, and many others " OFF \
+        "gpredict" "Satellite Tracking " OFF \
         "js8call" "js8call 2.20 JS8 for weak signal kbd-to-kbd messaging " OFF \
         "qsstv" "QSSTV 9.4.X for SSTV modes " OFF \
         "wsjtx" "WSJT-X 2.6.0 for FT4, FT8, JT4, JT9, JT65, Q65, MSK144 " OFF \
+        "xastir" "APRS Station Tracking and Reporting " OFF \
         3>&1 1>&2 2>&3)
     RET=$?
     if [ $RET -eq 1 ]; then
         $FUN = "NONE"
     fi
-    ##echo $FUN >> $SIGPI_INSTALLER
+
     IFS=' '     # space is set as delimiter
     read -ra ADDR <<< "$FUN"   # str is read into an array as tokens separated by IFS
     for i in "${ADDR[@]}"; do   # access each element of array
@@ -216,19 +222,39 @@ select_amateurradio() {
 }
 
 select_usefulapps() {
-    FUN=$(whiptail --title "SigPi Installer" --clear --checklist --separate-output \
-        "Useful Applications" 20 120 12 \
-        "HASviolet" "(RPi only) LoRa and FSK transceiver project " OFF \
+    FUN=$(whiptail --title "SIGpi Installer" --clear --checklist --separate-output \
+        "Chosse Additional and Advanced Applications" 20 120 12 \
+        "bettercap" "Swiss Army knife for 802.11, BLE, IPv4 and IPv6 reconnaissance" OFF \
         "cygnusrfi" "RFI) analysis tool, based on Python and GNU Radio Companion (GRC)" OFF \
-        "gpredict" "Satellite Tracking " OFF \
+        "HASviolet" "(RPi only) LoRa and FSK transceiver project " OFF \
 		"splat" "RF Signal Propagation, Loss, And Terrain analysis tool for 20 MHz to 20 GHz " OFF \
-        "xastir" "APRS Station Tracking and Reporting " OFF \
+        "srsran" "srsRAN, Open-source 4G/5G software radio suite " OFF \
+        "uniradhack" "Universal Radio Hacker. Offline wireless protocol investigation" OFF \
         3>&1 1>&2 2>&3)
     RET=$?
     if [ $RET -eq 1 ]; then
         $FUN = "NONE"
     fi
-    ##echo $FUN >> $SIGPI_INSTALLER
+
+    IFS=' '     # space is set as delimiter
+    read -ra ADDR <<< "$FUN"   # str is read into an array as tokens separated by IFS
+    for i in "${ADDR[@]}"; do   # access each element of array
+        echo $FUN >> $SIGPI_INSTALLER
+    done
+}
+
+select_nodeserver() {
+    FUN=$(whiptail --title "SIGpi Node Installer" --clear --checklist --separate-output \
+        "Choose which SDR server to run as a service" 20 120 12 \
+        "rtltcpsrv" "RTL-TCP Server " OFF \
+        "sdrangelsrv" "SDRangel Server" OFF \
+        "soapysdrsrv" "SoapySDR Server " OFF \
+		3>&1 1>&2 2>&3)
+    RET=$?
+    if [ $RET -eq 1 ]; then
+        $FUN = "NONE"
+    fi
+
     IFS=' '     # space is set as delimiter
     read -ra ADDR <<< "$FUN"   # str is read into an array as tokens separated by IFS
     for i in "${ADDR[@]}"; do   # access each element of array
@@ -247,10 +273,14 @@ touch $SIGPI_CONFIG
 mkdir $SIGPI_SOURCE
 cd $SIGPI_SOURCE
 
-# Node option invoked ?
+##
+## Node Install 
+##
+
 if [ "$1" = "node" ]; then
     calc_wt_size
-
+    select_devices
+    select_nodeserver
     TERM=ansi whiptail --title "SIGpi Node Install" --clear --msgbox "Ready to Install" 12 120
 
     echo -e "${SIGPI_BANNER_COLOR}"
@@ -268,54 +298,28 @@ if [ "$1" = "node" ]; then
 
     source $SIGPI_SCRIPTS/install_server_dependencies.sh
     source $SIGPI_SCRIPTS/install_core_devices.sh
-    source $SIGPI_PACKAGES/pkg_rtl_433 install
-    source $SIGPI_PACKAGES/pkg_dump1090 install
-    source $SIGPI_PACKAGES/pkg_multimon-ng install
-    source $SIGPI_PACKAGES/pkg_radiosonde install
-    source $SIGPI_PACKAGES/pkg_sigpi-node install $2
 
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   SIGpi Node Installation Complete !!"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   System needs to reboot for all changes to occur"
-    echo -e "${SIGPI_BANNER_COLOR} ##   Reboot will begin in 15 seconsds unless CTRL-C hit"
-    echo -e "${SIGPI_BANNER_RESET}"
-    sleep 15
-    sudo sync
-    sudo reboot
-    exit 0
+    # Install bladeRF
+    if grep bladerf "$SIGPI_INSTALLER"; then
+        source $SIGPI_PACKAGES/pkg_bladerf install
+    fi
+    # Install LimeSDR
+    if grep limesuite "$SIGPI_INSTALLER"; then
+        source $SIGPI_PACKAGES/pkg_limesuite install
+    fi
+    # Install Ettus UHD
+    if grep ettus "$SIGPI_INSTALLER"; then
+        source $SIGPI_PACKAGES/pkg_ettus install
+    fi
+    # Install SDRPlay
+    if grep sdrplay "$SIGPI_INSTALLER"; then
+        source $SIGPI_PACKAGES/pkg_sdrplay install
+    fi
+    # Install RFM95W (Adafruit RadioBonnet 900 MHz LoRa-FSK)
+    if grep rfm95w "$SIGPI_INSTALLER"; then
+        source $SIGPI_SCRIPTS/install_devices_rfm95w.sh
+    fi
 
-fi
-
-# base option invoked ?
-if [ "$1" = "base" ]; then
-    calc_wt_size
-
-    TERM=ansi whiptail --title "SIGpi Base Install" --clear --msgbox "Ready to Install" 12 120
-
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   System Update & Upgrade"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_RESET}"
-
-    sudo apt-get -y update
-    sudo apt-get -y upgrade
-
-    touch $SIGPI_CONFIG
-    echo "sigpi_desktop" >> $SIGPI_CONFIG
-    cd $SIGPI_SOURCE
-
-    # Install Core Dependencies
-    source $SIGPI_SCRIPTS/install_core_dependencies.sh
-    source $SIGPI_SCRIPTS/install_desktop-prep.sh
-    # Install Core Devices
-    source $SIGPI_SCRIPTS/install_core_devices.sh
-    # Install Libraries
-    source $SIGPI_SCRIPTS/install_libraries.sh
     # Install APTdec (NOAA APT)
     source $SIGPI_PACKAGES/pkg_aptdec install
     # Install NRSC5 (HD Radio)
@@ -344,14 +348,30 @@ if [ "$1" = "base" ]; then
     source $SIGPI_PACKAGES/pkg_rtl_433 install
     # Install Dump1090
     source $SIGPI_PACKAGES/pkg_dump1090 install
-    # Install MultiMon-NG
-    source $SIGPI_PACKAGES/pkg_multimon-ng install
-    # Install RadioSonde
-    source $SIGPI_PACKAGES/pkg_radiosonde install
+    
+    if [ "$2" = "rtltcpsrv" ]; then
+        sudo cp $SIGPI_SOURCE/scripts/sigpi-node_rtltcp.service /etc/systemd/system/sigpi-node.service
+        echo "sigpi-node_rtltcp" >> $SIGPI_CONFIG
+    elif [ "$2" = "sdrangelsrv" ]; then
+        sudo cp $SIGPI_SOURCE/scripts/sigpi-node_sdrangelsrv.service /etc/systemd/system/sigpi-node.service
+        echo "sigpi-node_sdrangelsrv" >> $SIGPI_CONFIG
+    elif [ "$2" = "soapysdrsrv" ]; then
+        sudo cp $SIGPI_SOURCE/scripts/sigpi-node_soapysdrsrv.service /etc/systemd/system/sigpi-node.service
+        echo "sigpi-node_soapysdrsrv" >> $SIGPI_CONFIG
+    else
+        echo -e "${SIGPI_BANNER_COLOR}"
+        echo -e "${SIGPI_BANNER_COLOR} ##  "
+        echo -e "${SIGPI_BANNER_COLOR} ##  ERROR: Unkown Server Type"
+        echo -e "${SIGPI_BANNER_COLOR} ##  "
+        echo -e "${SIGPI_BANNER_COLOR} ##  Your choices are rtltcp, sdrangelsrv, or soapysdrsrv"
+        echo -e "${SIGPI_BANNER_COLOR} ##  "
+        echo -e "${SIGPI_BANNER_RESET}"
+        exit 0
+    fi
 
     echo -e "${SIGPI_BANNER_COLOR}"
     echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   SIGpi Base Installation Complete !!"
+    echo -e "${SIGPI_BANNER_COLOR} ##   SIGpi Node Installation Complete !!"
     echo -e "${SIGPI_BANNER_COLOR} ##"
     echo -e "${SIGPI_BANNER_COLOR}"
     echo -e "${SIGPI_BANNER_COLOR} ##"
@@ -362,10 +382,89 @@ if [ "$1" = "base" ]; then
     sudo sync
     sudo reboot
     exit 0
+fi
+
+
+##
+## Base Install
+##
+
+if [ "$1" = "base" ]; then
+    calc_wt_size
+
+    TERM=ansi whiptail --title "SIGpi Base Install" --clear --msgbox "Ready to Install" 12 120
+
+    echo -e "${SIGPI_BANNER_COLOR}"
+    echo -e "${SIGPI_BANNER_COLOR} ##"
+    echo -e "${SIGPI_BANNER_COLOR} ##   System Update & Upgrade"
+    echo -e "${SIGPI_BANNER_COLOR} ##"
+    echo -e "${SIGPI_BANNER_RESET}"
+
+    sudo apt-get -y update
+    sudo apt-get -y upgrade
+
+    touch $SIGPI_CONFIG
+    echo "sigpi_desktop" >> $SIGPI_CONFIG
+    cd $SIGPI_SOURCE
+
+    # Install Core Dependencies
+    source $SIGPI_SCRIPTS/install_core_dependencies.sh
+    source $SIGPI_SCRIPTS/install_desktop-prep.sh
+    # Install Core Devices
+    source $SIGPI_SCRIPTS/install_core_devices.sh
+    # Install Libraries
+    source $SIGPI_SCRIPTS/install_libraries.sh
+    
+    # Install APTdec (NOAA APT)
+    source $SIGPI_PACKAGES/pkg_aptdec install
+    # Install NRSC5 (HD Radio)
+    source $SIGPI_PACKAGES/pkg_nrsc5 install
+    # Install cm256cc
+    source $SIGPI_PACKAGES/pkg_cm256cc install
+    # Install mbelib (P25 Phase)
+    source $SIGPI_PACKAGES/pkg_mbelib install
+    # Install SeriaDV (AMBE3000 chip serial control)
+    source $SIGPI_PACKAGES/pkg_serialdv install
+    # Install DSDcc (Digital Speech Decoder)
+    source $SIGPI_PACKAGES/pkg_dsdcc install
+    # Install Codec 2
+    source $SIGPI_PACKAGES/pkg_codec2 install
+    # Install Multimon-NG (POCSAG, FSK, AFSK, DTMF, X10)
+    source $SIGPI_PACKAGES/pkg_multimon-ng install
+    # Install Radiosonde (Atmospheric Telemetry)
+    source $SIGPI_PACKAGES/pkg_radiosonde install
+    # Install Ubertooth Tools
+    source $SIGPI_PACKAGES/pkg_ubertooth-tools install
+    # Install Direwolf (AFSK APRS)
+    source $SIGPI_PACKAGES/pkg_direwolf install
+    # Install Linpac (AX.25 Terminal
+    source $SIGPI_PACKAGES/pkg_linpac install
+    # Install RTL_433
+    source $SIGPI_PACKAGES/pkg_rtl_433 install
+    # Install Dump1090
+    source $SIGPI_PACKAGES/pkg_dump1090 install
+
+    echo -e "${SIGPI_BANNER_COLOR}"
+    echo -e "${SIGPI_BANNER_COLOR} ##"
+    echo -e "${SIGPI_BANNER_COLOR} ##   SIGpi Base Installation Complete !!"
+    echo -e "${SIGPI_BANNER_COLOR} ##"
+    echo -e "${SIGPI_BANNER_COLOR}"
+    echo -e "${SIGPI_BANNER_COLOR} ##"
+    echo -e "${SIGPI_BANNER_COLOR} ##   System needs to reboot for all changes to occur"
+    echo -e "${SIGPI_BANNER_COLOR} ##   Reboot will begin in 15 seconds unless CTRL-C hit"
+    echo -e "${SIGPI_BANNER_RESET}"
+    sleep 15
+    sudo sync
+    sudo reboot
+    exit 0
 
 fi
 
-# Otherwise we are Full install
+
+##
+## Full Install (Default)
+##
+
 calc_wt_size
 select_startscreen
 select_devices
@@ -397,17 +496,22 @@ source $SIGPI_SCRIPTS/install_desktop-prep.sh
 
 # Install Core Devices
 source $SIGPI_SCRIPTS/install_core_devices.sh
-
+# Install bladeRF
+if grep bladerf "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_bladerf install
+fi
 # Install LimeSDR
 if grep limesuite "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_limesuite install
 fi
-
 # Install Ettus UHD
 if grep ettus "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_ettus install
 fi
-
+# Install SDRPlay
+if grep sdrplay "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_sdrplay install
+fi
 # Install RFM95W (Adafruit RadioBonnet 900 MHz LoRa-FSK)
 if grep rfm95w "$SIGPI_INSTALLER"; then
     source $SIGPI_SCRIPTS/install_devices_rfm95w.sh
@@ -443,8 +547,11 @@ source $SIGPI_PACKAGES/pkg_linpac install
 source $SIGPI_PACKAGES/pkg_rtl_433 install
 # Install Dump1090
 source $SIGPI_PACKAGES/pkg_dump1090 install
-# Install GNUradio (3.10)
-source $SIGPI_PACKAGES/pkg_gnuradio install
+
+# GNU Radio
+if grep gnuradio "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_gnuradio-src install
+fi
 
 # CubicSDR
 if grep cubicsdr "$SIGPI_INSTALLER"; then
@@ -453,17 +560,22 @@ fi
 
 # SDRangel
 if grep sdrangel "$SIGPI_INSTALLER"; then
-    source $SIGPI_PACKAGES/pkg_sdrangel-deb install
+    source $SIGPI_PACKAGES/pkg_sdrangel install
 fi
 
 # SDR++
 if grep sdrpp "$SIGPI_INSTALLER"; then
-    source $SIGPI_PACKAGES/pkg_sdrpp-deb install
+    source $SIGPI_PACKAGES/pkg_sdrpp install
+fi
+
+# srsRAN
+if grep srsran "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_srsran install
 fi
 
 # HamLib
 if grep hamlib "$SIGPI_INSTALLER"; then
-    source $SIGPI_PACKAGES/pkg_hamlib-deb install
+    source $SIGPI_PACKAGES/pkg_hamlib install
 fi
 
 # Fldigi
@@ -471,7 +583,7 @@ if grep fldigi "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_fldigi install
 fi
 
- WSJT-X
+# WSJT-X
 if grep wsjtx "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_wsjtx install
 fi
@@ -501,6 +613,21 @@ if grep HASviolet "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_hasviolet install
 fi
 
+# Universal Radio Hacker
+if grep uniradhack "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_urh install
+fi
+
+# bettercap
+if grep bettercap "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_bettercap-src install
+fi
+
+# bettercapui
+#if grep bettercapui "$SIGPI_INSTALLER"; then
+#    source $SIGPI_PACKAGES/pkg_bettercapui-src install
+#fi
+
 # CygnusRFI
 if grep cygnusrfi "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_cygnusrfi install
@@ -515,15 +642,14 @@ source $SIGPI_PACKAGES/pkg_audacity install
 # Install PAVU 
 source $SIGPI_PACKAGES/pkg_pavucontrol install
 
-# splat
+# Install SPLAT
 if grep splat "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_splat install
 fi
 
-# DOSbox
+# Install DOSbox
 source $SIGPI_PACKAGES/pkg_dosbox install
-
-# SIGpi Menus
+# Install SIGpi Menus
 source $SIGPI_SCRIPTS/install_desktop-post.sh