+++ scripts/remove_desktop-prep.sh
@@ -24,23 +24,26 @@ sudo rm -rf /usr/local/bin/SIGpi
 
 # Remove SIGpi Background images
 ## What operating system are we?
-if [ "$SIGPI_OSNAME" = "Ubuntu 20.04.3 LTS" ]; then
+
+if [[ "$SIGPI_HWARCH" == "x86_64" ]]; then
     sudo rm -rf /usr/share/backgrounds/SIGpi*
-    sudo rm -rf /usr/share/backgrounds/SigPi*
+    sudo rm -rf /usr/share/backgrounds/SIGpi*
     # Change Background image
     gsettings set org.gnome.desktop.background picture-uri /usr/share/backgrounds/wallpaper.png
-else
+fi
+
+if [[ "$SIGPI_HWARCH" == "aarch64" ]]; then
     sudo rm -rf /usr/share/rpd-wallpaper/SIGpi*
-    sudo rm -rf /usr/share/rpd-wallpaper/SigPi*
+    sudo rm -rf /usr/share/rpd-wallpaper/SIGpi*
     # Change Background image
     pcmanfm --set-wallpaper=/usr/share/rpd-wallpaper/wallpaper.png
 fi
 
 # Remove SIGpi Desktop files from XDG Dirs
-sudo rm -rf $DESKTOP_DIRECTORY/SigPi.directory
-sudo rm -rf $DESKTOP_DIRECTORY/SigPishell.directory
-sudo rm -rf $DESKTOP_XDG_DIR/SigPi.menu
-sudo rm -rf $DESKTOP_XDG_DIR/SigPishell.menu
+sudo rm -rf $DESKTOP_DIRECTORY/SIGpi.directory
+sudo rm -rf $DESKTOP_DIRECTORY/SIGpishell.directory
+sudo rm -rf $DESKTOP_XDG_DIR/SIGpi.menu
+sudo rm -rf $DESKTOP_XDG_DIR/SIGpishell.menu
 
 echo -e "${SIGPI_BANNER_COLOR}"
 echo -e "${SIGPI_BANNER_COLOR} ##   Desktop Removal"