+++ scripts/run_urh.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+###
+### SIGpi URH Launch helper
+###
+### 
+
+cd $HOME/SIG/source/urh/src/urh
+./main.py