+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## Release 6.X: 2023-02-03
+- Rollback aarch64 to GNUradio 3.9.8 given unresolved issue with 3.10.5 compile
+- BladeRF support added
+- SDRPlay support added
+- SDRangel 7.9.0 release 
+- Universal Radio Hacker added
+- bettercap added
+- Beta of srsRAN
+
 ## Release 6.0.2: 2023-01-15
 - Maintenance Release
 - Ensure proper version of SDRangel is reflected (SDRangel 7.8.5)